+++ cocos/2d/CCLabelTextFormatter.cpp
@@ -47,18 +47,18 @@ bool LabelTextFormatter::multilineText(LabelTextFormatProtocol *theLabel)
         vector<unsigned short> last_word;
         last_word.reserve( stringLength );
         
-        unsigned int line = 1, i = 0;
+        long line = 1, i = 0;
         
         bool   isStartOfLine  = false, isStartOfWord = false;
         float  startOfLine = -1, startOfWord   = -1;
         
-        int skip = 0;
+        long skip = 0;
                 
-        int strLen = theLabel->getStringLenght();
+        long strLen = theLabel->getStringLenght();
         std::vector<LetterInfo>  *leterInfo = theLabel->getLettersInfo();
-        int tIndex = 0;
+        long tIndex = 0;
 
-        for (int j = 0; j < strLen; j++)
+        for (long j = 0; j < strLen; j++)
         {            
             LetterInfo* info = &leterInfo->at(j+skip);
 
@@ -264,7 +264,7 @@ bool LabelTextFormatter::alignText(LabelTextFormatProtocol *theLabel)
             
             if (shift != 0)
             {
-                for (unsigned j = 0; j < lineLength; ++j)
+                for (long j = 0; j < lineLength; ++j)
                 {
                     index = i + j + lineNumber;
                     if (index < 0) continue;
